Test suite for Transfer API method (#92)
