+++ __test__/integration/debt_token_api/debt_token_api.spec.ts
@@ -12,6 +12,8 @@ import {
     BALANCE_OF_SCENARIOS,
     OWNER_OF_SCENARIOS,
     EXISTS_SCENARIOS,
+    SUCCESSFUL_TRANSFER_SCENARIOS,
+    UNSUCCESSFUL_TRANSFER_SCENARIOS,
     SUCCESSFUL_TRANSFER_FROM_SCENARIOS,
     UNSUCCESSFUL_TRANSFER_FROM_SCENARIOS,
     SUCCESSFUL_APPROVE_SCENARIOS,
@@ -58,6 +60,16 @@ describe("Debt Token API (Integration Tests)", () => {
 
         describe("should fail", () => {
             UNSUCCESSFUL_APPROVE_SCENARIOS.forEach(scenarioRunner.testApproveScenario);
+        });
+    });
+
+    describe("#transfer", () => {
+        describe("should fail", () => {
+            UNSUCCESSFUL_TRANSFER_SCENARIOS.forEach(scenarioRunner.testTransferScenario);
+        });
+
+        describe("should succeed", () => {
+            SUCCESSFUL_TRANSFER_SCENARIOS.forEach(scenarioRunner.testTransferScenario);
         });
     });
