+++ __test__/integration/debt_token_api/runners/index.ts
@@ -110,5 +110,6 @@ export { BalanceOfScenarioRunner } from "./balance_of";
 export { OwnerOfScenarioRunner } from "./owner_of";
 export { ExistsScenarioRunner } from "./exists";
 export { TransferFromScenarioRunner } from "./transfer_from";
+export { TransferScenarioRunner } from "./transfer";
 export { ApproveScenarioRunner } from "./approve";
 export { GetApprovedScenarioRunner } from "./get_approved";
