+++ __test__/integration/debt_token_api/runners/transfer.ts
@@ -1,0 +1,10 @@
+// Base Runner
+import { TransferBaseScenarioRunner, TransferAPICallParameters } from "./transfer_base";
+
+export class TransferScenarioRunner extends TransferBaseScenarioRunner {
+    protected getAPICallPromise(params: TransferAPICallParameters): Promise<string> {
+        const { api, to, tokenID, txOptions } = params;
+
+        return api.transfer(to, tokenID, txOptions);
+    }
+}
