+++ __test__/integration/debt_token_api/scenarios/index.ts
@@ -2,7 +2,10 @@ export { DebtTokenScenario } from "./scenarios";
 export { BALANCE_OF_SCENARIOS } from "./balance_of";
 export { OWNER_OF_SCENARIOS } from "./owner_of";
 export { EXISTS_SCENARIOS } from "./exists";
-export { UNSUCCESSFUL_TRANSFER_FROM_SCENARIOS } from "./unsuccessful_transfer_from";
-export { SUCCESSFUL_TRANSFER_FROM_SCENARIOS } from "./successful_transfer_from";
+export { SUCCESSFUL_TRANSFER_SCENARIOS, UNSUCCESSFUL_TRANSFER_SCENARIOS } from "./transfer";
+export {
+    SUCCESSFUL_TRANSFER_FROM_SCENARIOS,
+    UNSUCCESSFUL_TRANSFER_FROM_SCENARIOS,
+} from "./transfer_from";
 export { SUCCESSFUL_APPROVE_SCENARIOS, UNSUCCESSFUL_APPROVE_SCENARIOS } from "./approve";
 export { GET_APPROVED_SCENARIOS } from "./get_approved";
